backend/requirements.txt
# AIML Recruitment Platform - Backend (Django + PostgreSQL)
Django>=4.2,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
pypdf>=4.0,<5.0
python-docx>=1.0,<2.0
scikit-learn>=1.3,<2.0
requests>=2.28,<3.0
Pillow>=10.0,<12.0
